new file was added

+++ 0x07-pointers_arrays_strings/2-strchr.c
@@ -7,7 +7,7 @@
  */
 char *_strchr(char *s, char c)
 {
-	unsigned int i = 0;
+	long long int i = 0;
 
 	while (s[i] != '\0')
 	{
